requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
pillow>=8.0.0
matplotlib>=3.3.2
tqdm>=4.50.0
scikit-image>=0.17.2
opencv-python>=4.4.0 